Renamed common package to util as it makes more sense
