+++ reactive-source-core/src/main/java/org/reactivesource/ReactiveDatasource.java
@@ -9,7 +9,7 @@ import com.google.common.annotations.VisibleForTesting;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.reactivesource.common.Assert.notNull;
+import static org.reactivesource.util.Assert.notNull;
 
 public class ReactiveDatasource<T> {
